basic script training a simple MLP to predict which loop a latent vector comes from
This experiment attempts to show indirectly, that the vector
space the recurrent block operates on is a number of distinct
subspaces. Especially for the first few loops, the MLP is able
to very reliably classify, and this accuracy gradually lowers.
Likely meaning past 8 loops, the latent space has converged to
a stable state.